// Employment/Frontend/src/app/interface/Model/ModelJob.class.ts
export class ModelJob{
    constructor(){
        this.Name = "";
    }
    Name: string;
    _id: string;
    Gender:number;
    CompanyId: string;
    Description: string;
    Skills:string;
    Benefits:string;
    Deadline: Date;
    JobFieldId: string;
    JobSubFieldId: string;
    ExperienceId: string;
    IndustryId: string;
    QualificationId: string;
    CountryId: string;
    CityId: string;
    Address: string;
    Type: number;
    Remuneration:string;
}